refactor(score): package 위치를 frame의 밖으로 옮김
