+++ src/main/java/bowling/domain/score/FinalScore.java
@@ -1,4 +1,4 @@
-package bowling.domain.frame.score;
+package bowling.domain.score;
 
 import bowling.domain.frame.Frame;
 import bowling.domain.pitch.Pitch;
@@ -6,9 +6,10 @@ import bowling.domain.pitch.Pitch;
 import java.util.Optional;
 
 public class FinalScore implements Score {
-    private static final int DEFAULT_MAX_ADD_NUMBER = 3;
-    private static final int ZERO = 0;
+    private static final int DEFAULT_MAX_ADD_NUMBER = 2;
     private static final int INIT_VALUE = 0;
+    private static final int ZERO = 0;
+    private static final int ONE = 1;
 
     private int maxAddNumber = DEFAULT_MAX_ADD_NUMBER;
     private int score;
@@ -34,9 +35,17 @@ public class FinalScore implements Score {
             return this;
         }
 
+        if(hasThirdChance(pitch)) {
+            maxAddNumber += ONE;
+        }
+
         maxAddNumber--;
         score += pitch.getPinCount();
         return this;
+    }
+
+    private boolean hasThirdChance(Pitch pitch) {
+        return pitch.isSpare() || pitch.isStrike();
     }
 
     private boolean isAddingDone() {
