+++ src/main/java/bowling/domain/score/FinishedScore.java
@@ -1,24 +1,19 @@
-package bowling.domain.frame.score;
+package bowling.domain.score;
 
 import bowling.domain.frame.Frame;
 import bowling.domain.pitch.Pitch;
 
-import java.util.Objects;
 import java.util.Optional;
 
-public class SpareScore implements Score {
+public class FinishedScore implements Score {
     private int score;
 
-    public SpareScore(int score) {
+    public FinishedScore(int score) {
         this.score = score;
     }
 
     @Override public Optional<Integer> calculateScore(Frame frame) {
-        if (Objects.isNull(frame)) {
-            return Optional.empty();
-        }
-        return frame.getPinCountForOnePitch()
-                .map(bonus -> bonus + score);
+        return calculateScore();
     }
 
     @Override public Optional<Integer> calculateScore() {
