+++ src/main/java/bowling/domain/score/SpareScore.java
@@ -1,19 +1,24 @@
-package bowling.domain.frame.score;
+package bowling.domain.score;
 
 import bowling.domain.frame.Frame;
 import bowling.domain.pitch.Pitch;
 
+import java.util.Objects;
 import java.util.Optional;
 
-public class FinishedScore implements Score {
+public class SpareScore implements Score {
     private int score;
 
-    public FinishedScore(int score) {
+    public SpareScore(int score) {
         this.score = score;
     }
 
     @Override public Optional<Integer> calculateScore(Frame frame) {
-        return calculateScore();
+        if (Objects.isNull(frame)) {
+            return Optional.empty();
+        }
+        return frame.getPinCountForOnePitch()
+                .map(bonus -> bonus + score);
     }
 
     @Override public Optional<Integer> calculateScore() {
